### 2019-05-30 #1484 @kyleniemeyer
Hi @haarburger, our build system can't find your paper (neither can I). Did you prepare a paper according to [our author guidelines](https://joss.readthedocs.io/en/latest/submitting.html)?
