### 2019-05-31 04cb947
Add debug mode
