### 2019-05-31 #1484 @haarburger
The paper was still in a non-master branch. I'll report back here after it is merged to master.
