+++ delira/data_loading/data_manager.py
@@ -2,8 +2,8 @@ import logging
 import numpy as np
 import typing
 import inspect
-from batchgenerators.dataloading import SlimDataLoaderBase, \
-    MultiThreadedAugmenter, SingleThreadedAugmenter
+from batchgenerators.dataloading import MultiThreadedAugmenter, \
+    SingleThreadedAugmenter
 from batchgenerators.transforms import AbstractTransform
 from .dataset import AbstractDataset, BaseCacheDataset, BaseLazyDataset, \
     ConcatDataset
@@ -17,17 +17,45 @@ logger = logging.getLogger(__name__)
 
 
 class Augmenter(object):
-    def __init__(self, data_loader: SlimDataLoaderBase, transforms,
-                 n_process_augmentation=None, num_cached_per_queue = 2,
+    """
+    Class wrapping ``MultiThreadedAugmentor`` and ``SingleThreadedAugmenter``
+    to provide a uniform API and to disable multiprocessing/multithreading
+    inside the dataloading pipeline
+
+    """
+    def __init__(self, data_loader: BaseDataLoader, transforms,
+                 n_process_augmentation=None, num_cached_per_queue=2,
                  seeds=None, **kwargs):
+        """
+
+        Parameters
+        ----------
+        data_loader : :class:`BaseDataLoader`
+            the dataloader providing the actual data
+        transforms : Callable or None
+            the transforms to use. Can be single callable or None
+        n_process_augmentation : int
+            the number of processes to use for augmentation (only necessary if
+            not in debug mode)
+        num_cached_per_queue : int
+            the number of samples to cache per queue (only necessary if not in
+            debug mode)
+        seeds : int or list
+            the seeds for each process (only necessary if not in debug mode)
+        **kwargs :
+            additional keyword arguments
+        """
+        # don't use multiprocessing in debug mode
         if get_current_debug_mode():
             augmenter = SingleThreadedAugmenter(data_loader, transforms)
 
         else:
             assert isinstance(n_process_augmentation, int)
+            # no seeds are given -> use default seed of 1
             if seeds is None:
                 seeds = 1
 
+            # only an int is gien as seed -> replicate it for each process
             if isinstance(seeds, int):
                 seeds = [seeds] * n_process_augmentation
             augmenter = MultiThreadedAugmenter(
@@ -41,17 +69,73 @@ class Augmenter(object):
 
     @property
     def __iter__(self):
+        """
+        Property returning the augmenters ``__iter__``
+
+        Returns
+        -------
+        Callable
+            the augmenters ``__iter__``
+        """
         return self._augmenter.__iter__
 
     @property
     def __next__(self):
+        """
+        Property returning the augmenters ``__next__``
+
+        Returns
+        -------
+        Callable
+            the augmenters ``__next__``
+        """
         return self._augmenter.__next__
+
+    @property
+    def next(self):
+        """
+        Property returning the augmenters ``next``
+
+        Returns
+        -------
+        Callable
+            the augmenters ``next``
+        """
+        return self._augmenter.next
 
     @staticmethod
     def __identity_fn(*args, **kwargs):
+        """
+        Helper function accepting arbitrary args and kwargs and returning
+        without doing anything
+
+        Parameters
+        ----------
+        *args
+            keyword arguments
+        **kwargs
+            positional arguments
+
+        """
         return
 
     def _fn_checker(self, function_name):
+        """
+        Checks if the internal augmenter has a given attribute and returns it.
+        Otherwise it returns ``__identity_fn``
+
+        Parameters
+        ----------
+        function_name : str
+            the function name to check for
+
+        Returns
+        -------
+        Callable
+            either the function corresponding to the given function name or
+            ``__identity_fn``
+
+        """
         # same as:
         # if hasattr(self._augmenter, function_name):
[…]
