### 2019-06-01 7c2b81a
Add kwargs in predictor for prepare_batch_fn
