+++ delira/data_loading/data_manager.py
@@ -3,7 +3,7 @@ import numpy as np
 import typing
 import inspect
 from batchgenerators.dataloading import MultiThreadedAugmenter, \
-    SingleThreadedAugmenter
+    SingleThreadedAugmenter, SlimDataLoaderBase
 from batchgenerators.transforms import AbstractTransform
 from .dataset import AbstractDataset, BaseCacheDataset, BaseLazyDataset, \
     ConcatDataset
@@ -23,6 +23,7 @@ class Augmenter(object):
     inside the dataloading pipeline
 
     """
+
     def __init__(self, data_loader: BaseDataLoader, transforms,
                  n_process_augmentation=None, num_cached_per_queue=2,
                  seeds=None, **kwargs):
@@ -362,12 +363,12 @@ class BaseDataManager(object):
         return Augmenter(data_loader, self.transforms,
                          self.n_process_augmentation,
                          num_cached_per_queue=2,
-                         seeds=self.n_process_augmentation*[seed])
+                         seeds=self.n_process_augmentation * [seed])
 
     def get_subset(self, indices):
         """
         Returns a Subset of the current datamanager based on given indices
-        
+
         Parameters
         ----------
         indices : iterable
@@ -377,7 +378,7 @@ class BaseDataManager(object):
         -------
         :class:`BaseDataManager`
             manager containing the subset
-        
+
         """
 
         subset_kwargs = {
@@ -397,7 +398,7 @@ class BaseDataManager(object):
         """
         Updates internal state and therfore the behavior from dict.
         If a key is not specified, the old attribute value will be used
-        
+
         Parameters
         ----------
         new_state : dict
@@ -418,7 +419,7 @@ class BaseDataManager(object):
         ------
         KeyError
             Invalid keys are specified
-        
+
         """
 
         # update batch_size if specified
@@ -458,7 +459,7 @@ class BaseDataManager(object):
         **kwargs :
             keyword arguments for 
             ``sklearn.model_selection.train_test_split``
-        
+
         """
 
         trainset, valset = self.dataset.train_test_split(*args, **kwargs)
@@ -483,7 +484,7 @@ class BaseDataManager(object):
     def batch_size(self):
         """
         Property to access the batchsize
-        
+
         Returns
         -------
         int
@@ -496,13 +497,13 @@ class BaseDataManager(object):
     def batch_size(self, new_batch_size):
         """
         Setter for current batchsize, casts to int before setting the attribute
-        
+
         Parameters
         ----------
         new_batch_size : int, Any
             the new batchsize; should be int but can be of any type that can be 
             casted to an int
-        
+
         """
 
         self._batch_size = int(new_batch_size)
@@ -511,7 +512,7 @@ class BaseDataManager(object):
     def n_process_augmentation(self):
         """
         Property to access the number of augmentation processes
-        
+
         Returns
         -------
         int
@@ -525,13 +526,13 @@ class BaseDataManager(object):
         """
         Setter for number of augmentation processes, casts to int before setting
         the attribute
-        
+
         Parameters
         ----------
         new_process_number : int, Any
             new number of augmentation processes; should be int but can be of 
             any type that can be casted to an int
-        
+
         """
 
         self._n_process_augmentation = int(new_process_number)
@@ -540,7 +541,7 @@ class BaseDataManager(object):
     def transforms(self):
         """
         Property to access the current data transforms
-        
+
         Returns
         -------
         None, ``AbstractTransform``
@@ -555,12 +556,12 @@ class BaseDataManager(object):
         """
         Setter for data transforms, assert if transforms are of valid type 
         (either None or instance of ``AbstractTransform``)
-        
+
         Parameters
         ----------
         new_transforms : None, ``AbstractTransform``
[…]
