### 2019-06-01 bf07a4d
Add @mibaumgartner to core developers
