+++ AUTHORS.rst
@@ -4,12 +4,13 @@ Authors
 
 **Core Development Team:**
 
-- Justus Schock: `LinkedIn <https://www.linkedin.com/in/justus-schock/>`_ | `Google Scholar <https://scholar.google.de/citations?hl=de&user=KYf-ZHoAAAAJ>`_ | `E-Mail <mailto:justus.schock@rwth-aachen.de>`_
-- Oliver Rippel: `LinkedIn <https://www.linkedin.com/in/oliver-rippel-70361113a/>`_ | `Google Scholar <https://scholar.google.de/citations?user=DaTF8RsAAAAJ&hl=de>`_
-- Christoph Haarburger: `LinkedIn <https://www.linkedin.com/in/chaarburger/>`_ | `Google Scholar <https://scholar.google.de/citations?user=Lb8DcccAAAAJ&hl=de>`_ 
+- Justus Schock: `GitHub <https://github.com/justusschock>`_ | `LinkedIn <https://www.linkedin.com/in/justus-schock/>`_ | `Google Scholar <https://scholar.google.de/citations?hl=de&user=KYf-ZHoAAAAJ>`_ | `E-Mail <mailto:justus.schock@rwth-aachen.de>`_
+- Michael Baumgartner: `GitHub <https://github.com/mibaumgartner>`_ | `LinkedIn <https://www.linkedin.com/in/michael-baumgartner-/>`_
+- Oliver Rippel: `GitHub <https://github.com/ORippler>`_ | `LinkedIn <https://www.linkedin.com/in/oliver-rippel-70361113a/>`_ | `Google Scholar <https://scholar.google.de/citations?user=DaTF8RsAAAAJ&hl=de>`_
+- Christoph Haarburger: `GitHub <https://github.com/haarburger>`_ | `LinkedIn <https://www.linkedin.com/in/chaarburger/>`_ | `Google Scholar <https://scholar.google.de/citations?user=Lb8DcccAAAAJ&hl=de>`_ 
 
 **Contributions:**
 
 - Nicolas Horst
-- Michael Baumgartner
+
 - Alexander Moriz
