### 2019-06-01 f6d0888
Update CONTRIBUTING.md
