### 2019-06-03 0d9a89d
Add docstrings
