### 2019-06-03 47e2556
Convert batchdict back to numpy for metric calculation and introduce lazy  behavior
