### 2019-06-03 b510ffb
fix generator behavior
