### 2019-06-03 13d3adb
make experiment.test return the first generator item
