### 2019-06-03 8769912
Update .travis.yml
