+++ .travis.yml
@@ -1,8 +1,5 @@
 language: python
 
-services:
-    - docker
-
 matrix:
     include:
         - name: "Unittests Python 3.5"
@@ -40,17 +37,17 @@ install:
 # command to run tests
 script:
     # run tests or stylechecks
-    - if [[ "$TEST_TYPE" == "unittests" ]]; then
-          bash scripts/ci/run_tests.sh;
+    if [[ "$TEST_TYPE" == "unittests" ]]; then
+        bash scripts/ci/run_tests.sh;
       
           # build docs if python == 3.7
-          if [[ "$TRAVIS_PYTHON_VERSION" == "3.7" ]]; then
-              bash scripts/ci/build_docs.sh;
-          fi
+        if [[ "$TRAVIS_PYTHON_VERSION" == "3.7" ]]; then
+            bash scripts/ci/build_docs.sh;
+        fi
 
-      else
-          bash scripts/ci/install_before_style_check.sh;
-      fi
+    else
+        bash scripts/ci/install_before_style_check.sh;
+    fi
 
 before_deploy:
     - cd $TRAVIS_BUILD_DIR
