### 2019-06-03 dd17002
Correct script to run style checks
