### 2019-06-03 7fab627
Update run_style_checks.sh
