### 2019-06-03 07fc243
Update install_before_style_check.sh
