+++ scripts/ci/install_before_style_check.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 
-pip install pep8;
+pip install pycodestyle;
 pip install autopep8;
