### 2019-06-03 51d3bed
Update run_style_checks.sh
