### 2019-06-03 eb0fe22
Update run_style_checks.sh
