### 2019-06-03 c558096
Update run_style_checks.sh
