### 2019-06-03 9aacdab
Update setup.cfg
