+++ setup.cfg
@@ -1,2 +1,2 @@
-[pep8]
+[pycodestyle]
 exclude = .eggs,*.egg,build,docs,.git
