### 2019-06-03 9ae6d87
Start making codebase pep8 compatible
