### 2019-06-03 #1484 @mbobra
@sealhuang and @fernandoandreotti Are you available to review this submission for JOSS?
