### 2019-06-03 d123c1e
add kwargs to overwritten predict_data_mgr functions
