+++ delira/training/experiment.py
@@ -390,7 +390,8 @@ class BaseExperiment(TrixiExperiment):
 
         # return first item of generator
         return next(predictor.predict_data_mgr(test_data, 1, metrics,
-                                               metric_keys, verbose))
+                                               metric_keys, verbose,
+                                               lazy_gen=False))
 
     def kfold(self, data: BaseDataManager, metrics: dict, num_epochs=None,
               num_splits=None, shuffle=False, random_seed=None,
