+++ delira/training/pytorch_trainer.py
@@ -378,7 +378,7 @@ if "TORCH" in get_backends():
                                                verbose=verbose)
 
         def predict_data_mgr(self, datamgr, batchsize=None, metrics={},
-                             metric_keys={}, verbose=False):
+                             metric_keys={}, verbose=False, **kwargs):
             """
             Defines a routine to predict data obtained from a batchgenerator
 
@@ -396,6 +396,8 @@ if "TORCH" in get_backends():
                 the ``batch_dict`` items to use for metric calculation
             verbose : bool
                 whether to show a progress-bar or not, default: False
+            **kwargs :
+                additional keyword arguments
 
             Returns
             -------
@@ -408,7 +410,7 @@ if "TORCH" in get_backends():
             self.module.eval()
 
             return super().predict_data_mgr(datamgr, batchsize, metrics,
-                                            metric_keys, verbose)
+                                            metric_keys, verbose, **kwargs)
 
         def save_state(self, file_name, epoch, **kwargs):
             """
