### 2019-06-03 eb6a5f9
make delira pep8 compatible
