+++ tests/models/test_models_torch.py
@@ -18,8 +18,8 @@ class TorchModelTest(unittest.TestCase):
             UNet3dPyTorch, ClassificationNetworkBasePyTorch, \
             VGG3DClassificationNetworkPyTorch, \
             GenerativeAdversarialNetworkBasePyTorch
-        from delira.training.train_utils import create_optims_default_pytorch, \
-            create_optims_gan_default_pytorch
+        from delira.training.train_utils import \
+            create_optims_default_pytorch, create_optims_gan_default_pytorch
         from delira.utils.context_managers import DefaultOptimWrapperTorch
         import torch
 
