+++ delira/utils/__init__.py
@@ -1,6 +1,6 @@
-from .imageops import bounding_box, calculate_origin_offset, max_energy_slice, \
-    sitk_new_blank_image, sitk_resample_to_image, sitk_resample_to_shape, \
-    sitk_resample_to_spacing
+from .imageops import bounding_box, calculate_origin_offset, \
+    max_energy_slice, sitk_new_blank_image, sitk_resample_to_image, \
+    sitk_resample_to_shape, sitk_resample_to_spacing
 
 from .path import subdirs
 from .time import now
