### 2019-06-03 f4bdd2c
final pep8 fixes
