### 2019-06-03 8035035
fix script in ci/cd and write out list comprehensions
