+++ tests/models/test_models_torch.py
@@ -1,10 +1,11 @@
-from delira import get_backends
-import numpy as np
 import time
-import gc
-import sys
-from psutil import virtual_memory
 import unittest
+
+import gc
+import numpy as np
+from psutil import virtual_memory
+
+from delira import get_backends
 
 
 class TorchModelTest(unittest.TestCase):
@@ -14,7 +15,7 @@ class TorchModelTest(unittest.TestCase):
     @unittest.skipIf("TORCH" not in get_backends(),
                      reason="torch backend not installed")
     def test_pytorch_model_default(self):
-        from delira.models import AbstractPyTorchNetwork, UNet2dPyTorch, \
+        from delira.models import UNet2dPyTorch, \
             UNet3dPyTorch, ClassificationNetworkBasePyTorch, \
             VGG3DClassificationNetworkPyTorch, \
             GenerativeAdversarialNetworkBasePyTorch
