+++ delira/data_loading/sampler/abstract_sampler.py
@@ -1,4 +1,5 @@
 from abc import abstractmethod
+
 from ..dataset import AbstractDataset
 
 
