+++ delira/data_loading/nii.py
@@ -1,10 +1,13 @@
-import logging
-import SimpleITK as sitk
-import numpy as np
 import json
+import logging
 import os
 from abc import abstractmethod
+
+import SimpleITK as sitk
+import numpy as np
+
 from delira.utils.decorators import make_deprecated
+
 logger = logging.getLogger(__name__)
 
 
