+++ delira/models/classification/classification_network.py
@@ -1,5 +1,7 @@
-from delira import get_backends
 import logging
+
+from delira import get_backends
+
 file_logger = logging.getLogger(__name__)
 
 
