+++ delira/training/callbacks/pytorch_schedulers.py
@@ -1,5 +1,5 @@
-from .abstract_callback import AbstractCallback
 from delira import get_backends
+from .abstract_callback import AbstractCallback
 
 if 'TORCH' in get_backends():
     from torch.optim.lr_scheduler import ReduceLROnPlateau, \
