+++ delira/utils/__init__.py
@@ -1,8 +1,6 @@
+from .config import LookupConfig
 from .imageops import bounding_box, calculate_origin_offset, \
     max_energy_slice, sitk_new_blank_image, sitk_resample_to_image, \
     sitk_resample_to_shape, sitk_resample_to_spacing
-
 from .path import subdirs
 from .time import now
-
-from .config import LookupConfig
