### 2019-06-04 3d0c742
Update .travis.yml
