### 2019-06-04 a026551
Update .travis.yml
