+++ delira/models/classification/classification_network_3D.py
@@ -1,11 +1,10 @@
+from delira.utils.decorators import make_deprecated
 import logging
 
 from delira import get_backends
 
 file_logger = logging.getLogger(__name__)
 
-
-from delira.utils.decorators import make_deprecated
 
 if "TORCH" in get_backends():
     import torch.nn as nn

[…]
