### 2019-06-04 b63587c
Merge GAN
