### 2019-06-04 2e264f5
Add style fixes and common function to search for previous checkpoints
