+++ delira/training/pytorch_trainer.py
@@ -227,24 +227,10 @@ if "TORCH" in get_backends():
 
             # Load latest epoch file if available
             if os.path.isdir(self.save_path):
-                # check all files in directory starting with "checkpoint" and
-                # not ending with "_best.pth"
-                files = [x for x in os.listdir(self.save_path)
-                         if os.path.isfile(os.path.join(self.save_path, x))
-                         and x.startswith("checkpoint")
-                         and not (x.endswith("_best.pth")
-                                  or x.endswith("_best.pt"))]
+                latest_state_path, latest_epoch = self._search_for_prev_state(
+                    self.save_path, [".pt", ".pth"])
 
-                # if list is not empty: load previous state
-                if files:
-
-                    latest_epoch = max([
-                        int(x.rsplit("_", 1)[-1].rsplit(".", 1)[0])
-                        for x in files])
-
-                    latest_state_path = os.path.join(self.save_path,
-                                                     "checkpoint_epoch_%d.pth"
-                                                     % latest_epoch)
+                if latest_state_path is not None:
 
                     # if pth file does not exist, load pt file instead
                     if not os.path.isfile(latest_state_path):
@@ -255,7 +241,7 @@ if "TORCH" in get_backends():
                     try:
                         self.update_state(latest_state_path)
                     except KeyError:
-                        logger.warn("Previous State could not be loaded, \
+                        logger.warning("Previous State could not be loaded, \
                                     although it exists.Training will be \
                                     restarted")
 
