### 2019-06-04 bcee731
Remove TrixiExperiment as Experiment Baseclass
