### 2019-06-04 ac1c6e9
correct indent in tf trainer
