+++ delira/training/tf_trainer.py
@@ -205,8 +205,8 @@ class TfNetworkTrainer(BaseNetworkTrainer):
                 logger.info("Attempting to load state from previous \
                             training from %s" % latest_state_path)
 
-            self.update_state(latest_state_path)
-            self.start_epoch = latest_epoch
+                self.update_state(latest_state_path)
+                self.start_epoch = latest_epoch
 
     def _at_training_end(self):
         """
