+++ delira/models/__init__.py
@@ -13,4 +13,3 @@ if "TORCH" in get_backends():
 if "TF" in get_backends():
     from .abstract_network import AbstractTfNetwork, AbstractTfEagerNetwork
     from .classification import ClassificationNetworkBaseTf
-
