+++ delira/data_loading/sampler/random_sampler.py
@@ -73,6 +73,7 @@ class RandomSamplerNoReplacement(RandomSampler):
     """
     Implements Random Sampling Without Replacement from whole Dataset
     """
+
     def __init__(self, indices):
         """
 
