+++ delira/models/model_utils.py
@@ -23,7 +23,7 @@ def scale_loss(loss,
     else:
         with amp.scale_loss(loss=loss, optimizers=optimizers,
                             loss_id=loss_id, model=model,
-                            delay_unscale=delay_unscale, 
+                            delay_unscale=delay_unscale,
                             **kwargs) as _loss:
 
             yield _loss
