### 2019-06-04 bff8831
Update .travis.yml
