### 2019-06-04 955003a
Make installation dependent on environment variable
