### 2019-06-04 be36b63
Update install_before_tests.sh
