### 2019-06-04 8c1002a
Update .travis.yml
