+++ setup.py
@@ -54,7 +54,7 @@ def parse_all_requirements(backend_requirement_dict: dict):
 
 requirement_files = {"base": "requirements.txt",
                      "torch": "requirements_extra_torch.txt",
-                     "torchscript": "requirements_extra.txt",
+                     "torchscript": "requirements_extra_torch.txt",
                      "tf": "requirements_extra_tf.txt"
                     }
 
