### 2019-06-04 204ea00
Fix gpu device id
