### 2019-06-04 11dbf4a
Fix gpu device id
