### 2019-06-04 70cb4b6
Add kwargs to tf_eager trainer
