+++ setup.py
@@ -55,7 +55,8 @@ def parse_all_requirements(backend_requirement_dict: dict):
 requirement_files = {"base": "requirements.txt",
                      "torch": "requirements_extra_torch.txt",
                      "torchscript": "requirements_extra_torch.txt",
-                     "tf": "requirements_extra_tf.txt"
+                     "tf": "requirements_extra_tf.txt",
+                     "tf_eager": "requirements_extra_tf"
                      }
 
 
