### 2019-06-05 4ac64e1
add backend specific funtion as default argument
