### 2019-06-05 924ce53
add backend specific funtion as default argument
