### 2019-06-05 e0ef923
Add scikitlearn test workers
