+++ delira/training/__init__.py
@@ -18,4 +18,3 @@ if "TF" in get_backends():
 if "SKLEARN" in get_backends():
     from .sklearn_trainer import SklearnEstimatorTrainer
     from .experiment import SkLearnExperiment
-
