+++ delira/io/__init__.py
@@ -17,5 +17,3 @@ if "TF" in get_backends():
 if "SKLEARN" in get_backends():
     from .sklearn import load_checkpoint as sklearn_load_checkpoint
     from .sklearn import save_checkpoint as sklearn_save_checkpoint
-
-
