### 2019-06-05 c65fafd
Add installation of chainer dependencies
