### 2019-06-05 59e455a
make tests skipping if chainer backend not installed
