### 2019-06-05 49e07ae
make labels 1d by default for scikit learn backend
