### 2019-06-05 c0634c7
fix testcases for chainer backend
