### 2019-06-05 6b9ac0d
shallow copy
Make shallow copy of batchdict to retain keys, which might get popped in `prepare_batch`
