### 2019-06-05 13fe994
Add kwargs to ChainerNetworkTrainer.predict_data_mgr
