+++ delira/training/chainer_trainer.py
@@ -388,7 +388,7 @@ if "CHAINER" in get_backends():
                                                verbose=verbose)
 
         def predict_data_mgr(self, datamgr, batchsize=None, metrics={},
-                             metric_keys={}, verbose=False):
+                             metric_keys={}, verbose=False, **kwargs):
             """
             Defines a routine to predict data obtained from a batchgenerator
 
@@ -406,6 +406,8 @@ if "CHAINER" in get_backends():
                 the ``batch_dict`` items to use for metric calculation
             verbose : bool
                 whether to show a progress-bar or not, default: False
+            **kwargs :
+                additional keyword arguments
 
             Returns
             -------
@@ -418,7 +420,7 @@ if "CHAINER" in get_backends():
             chainer.global_config.train = False
 
             return super().predict_data_mgr(datamgr, batchsize, metrics,
-                                            metric_keys, verbose)
+                                            metric_keys, verbose, **kwargs)
 
         def save_state(self, file_name, epoch, **kwargs):
             """
