+++ delira/models/backends/tf_eager/__init__.py
@@ -1,0 +1,1 @@
+from .abstract_network import AbstractTfEagerNetwork
