+++ delira/training/backends/torchscript/trainer.py
@@ -1,0 +1,235 @@
+import logging
+import os
+from functools import partial
+import warnings
+
+import torch
+from batchgenerators.dataloading import MultiThreadedAugmenter
+
+from delira.io.torch import load_checkpoint_torchscript, \
+    save_checkpoint_torchscript
+from delira.models import AbstractTorchScriptNetwork
+
+from delira.training.base_trainer import BaseNetworkTrainer
+from ..torch.trainer import PyTorchNetworkTrainer
+
+from .utils import convert_to_numpy, create_optims_default
+
+
+logger = logging.getLogger(__name__)
+
+
+class TorchScriptNetworkTrainer(PyTorchNetworkTrainer):
+    def __init__(self,
+                 network: AbstractTorchScriptNetwork,
+                 save_path: str,
+                 key_mapping,
+                 losses=None,
+                 optimizer_cls=None,
+                 optimizer_params={},
+                 train_metrics={},
+                 val_metrics={},
+                 lr_scheduler_cls=None,
+                 lr_scheduler_params={},
+                 gpu_ids=[],
+                 save_freq=1,
+                 optim_fn=create_optims_default,
+                 logging_type="tensorboardx",
+                 logging_kwargs={},
+                 fold=0,
+                 callbacks=[],
+                 start_epoch=1,
+                 metric_keys=None,
+                 convert_batch_to_npy_fn=convert_to_numpy,
+                 criterions=None,
+                 val_freq=1,
+                 **kwargs):
+        """
+
+        Parameters
+        ----------
+        network : :class:`AbstractPyTorchJITNetwork`
+            the network to train
+        save_path : str
+            path to save networks to
+        key_mapping : dict
+            a dictionary containing the mapping from the ``data_dict`` to
+            the actual model's inputs.
+            E.g. if a model accepts one input named 'x' and the data_dict
+            contains one entry named 'data' this argument would have to
+            be ``{'x': 'data'}``
+        losses : dict
+            dictionary containing the training losses
+        optimizer_cls : subclass of tf.train.Optimizer
+            optimizer class implementing the optimization algorithm of
+            choice
+        optimizer_params : dict
+            keyword arguments passed to optimizer during construction
+        train_metrics : dict, optional
+            metrics, which will be evaluated during train phase
+            (should work on framework's tensor types)
+        val_metrics : dict, optional
+            metrics, which will be evaluated during test phase
+            (should work on numpy arrays)
+        lr_scheduler_cls : Any
+            learning rate schedule class: must implement step() method
+        lr_scheduler_params : dict
+            keyword arguments passed to lr scheduler during construction
+        gpu_ids : list
+            list containing ids of GPUs to use; if empty: use cpu instead
+            Currently ``torch.jit`` only supports single GPU-Training,
+            thus only the first GPU will be used if multiple GPUs are
+            passed
+        save_freq : int
+            integer specifying how often to save the current model's state.
+            State is saved every state_freq epochs
+        optim_fn : function
+            creates a dictionary containing all necessary optimizers
+        logging_type : str or callable
+            the type of logging. If string: it must be one of
+            ["visdom", "tensorboardx"]
+            If callable: it must be a logging handler class
+        logging_kwargs : dict
+            dictionary containing all logging keyword arguments
+        fold : int
+            current cross validation fold (0 per default)
+        callbacks : list
+            initial callbacks to register
+        start_epoch : int
+            epoch to start training at
+        metric_keys : dict
+            dict specifying which batch_dict entry to use for which metric
+            as target; default: None, which will result in key "label" for
+            all metrics
+        convert_batch_to_npy_fn : type, optional
+            function converting a batch-tensor to numpy, per default this
[…]
