+++ delira/training/backends/torch/experiment.py
@@ -1,0 +1,264 @@
+from functools import partial
+import typing
+
+import torch
+
+from delira.models import AbstractPyTorchNetwork
+from delira.data_loading import BaseDataManager
+
+from delira.training.experiment import BaseExperiment
+from delira.training.parameters import Parameters
+
+from .trainer import PyTorchNetworkTrainer
+from .utils import create_optims_default, convert_to_numpy
+
+
+class PyTorchExperiment(BaseExperiment):
+    def __init__(self,
+                 params: typing.Union[str, Parameters],
+                 model_cls: AbstractPyTorchNetwork,
+                 n_epochs=None,
+                 name=None,
+                 save_path=None,
+                 key_mapping=None,
+                 val_score_key=None,
+                 optim_builder=create_optims_default,
+                 checkpoint_freq=1,
+                 trainer_cls=PyTorchNetworkTrainer,
+                 **kwargs):
+        """
+
+        Parameters
+        ----------
+        params : :class:`Parameters` or str
+            the training parameters, if string is passed,
+            it is treated as a path to a pickle file, where the
+            parameters are loaded from
+        model_cls : Subclass of :class:`AbstractPyTorchNetwork`
+            the class implementing the model to train
+        n_epochs : int or None
+            the number of epochs to train, if None: can be specified later
+            during actual training
+        name : str or None
+            the Experiment's name
+        save_path : str or None
+            the path to save the results and checkpoints to.
+            if None: Current working directory will be used
+        key_mapping : dict
+            mapping between data_dict and model inputs (necessary for
+            prediction with :class:`Predictor`-API), if no keymapping is
+            given, a default key_mapping of {"x": "data"} will be used here
+        val_score_key : str or None
+            key defining which metric to use for validation (determining
+            best model and scheduling lr); if None: No validation-based
+            operations will be done (model might still get validated,
+            but validation metrics can only be logged and not used further)
+        optim_builder : function
+            Function returning a dict of backend-specific optimizers.
+            defaults to :func:`create_optims_default_pytorch`
+        checkpoint_freq : int
+            frequency of saving checkpoints (1 denotes saving every epoch,
+            2 denotes saving every second epoch etc.); default: 1
+        trainer_cls : subclass of :class:`PyTorchNetworkTrainer`
+            the trainer class to use for training the model, defaults to
+            :class:`PyTorchNetworkTrainer`
+        **kwargs :
+            additional keyword arguments
+
+        """
+
+        if key_mapping is None:
+            key_mapping = {"x": "data"}
+        super().__init__(params=params, model_cls=model_cls,
+                         n_epochs=n_epochs, name=name, save_path=save_path,
+                         key_mapping=key_mapping,
+                         val_score_key=val_score_key,
+                         optim_builder=optim_builder,
+                         checkpoint_freq=checkpoint_freq,
+                         trainer_cls=trainer_cls,
+                         **kwargs)
+
+    def kfold(self, data: BaseDataManager, metrics: dict, num_epochs=None,
+              num_splits=None, shuffle=False, random_seed=None,
+              split_type="random", val_split=0.2, label_key="label",
+              train_kwargs: dict = None, test_kwargs: dict = None,
+              metric_keys: dict = None, params=None, verbose=False,
+              **kwargs):
+        """
+        Performs a k-Fold cross-validation
+
+        Parameters
+        ----------
+        data : :class:`BaseDataManager`
+            the data to use for training(, validation) and testing. Will be
+            split based on ``split_type`` and ``val_split``
+        metrics : dict
+            dictionary containing the metrics to evaluate during k-fold
+        num_epochs : int or None
+            number of epochs to train (if not given, will either be
+            extracted from ``params``, ``self.parms`` or ``self.n_epochs``)
+        num_splits : int or None
[…]
