+++ delira/models/backends/torchscript/__init__.py
@@ -1,0 +1,1 @@
+from .abstract_network import AbstractTorchScriptNetwork

[…]
