+++ delira/training/backends/torch/experiment.py
@@ -6,7 +6,7 @@ import torch
 from delira.models import AbstractPyTorchNetwork
 from delira.data_loading import BaseDataManager
 
-from delira.training.experiment import BaseExperiment
+from delira.training.base_experiment import BaseExperiment
 from delira.training.parameters import Parameters
 
 from .trainer import PyTorchNetworkTrainer
