+++ delira/training/backends/tf_eager/__init__.py
@@ -1,0 +1,5 @@
+from .experiment import TfEagerExperiment
+from .trainer import TfEagerNetworkTrainer
+from .utils import convert_to_numpy as convert_tfeager_to_numpy, \
+    switch_tf_execution_mode, create_optims_default as \
+    create_tfeager_optims_default
