+++ delira/training/backends/sklearn/experiment.py
@@ -6,7 +6,7 @@ from sklearn.base import BaseEstimator
 
 from delira.models import SklearnEstimator
 
-from delira.training.experiment import BaseExperiment
+from delira.training.base_experiment import BaseExperiment
 from delira.training.parameters import Parameters
 
 from .trainer import SklearnEstimatorTrainer
