### 2019-06-05 608c221
Start Refactoring Tests
