+++ tests/models/test_abstract_models.py
@@ -1,0 +1,206 @@
+import unittest
+import numpy as np
+from delira import get_backends
+
+
+class TestAbstractModels(unittest.TestCase):
+
+    @staticmethod
+    def _setup_torch(*args):
+        import torch
+        from delira.models import AbstractPyTorchNetwork
+
+        class Model(AbstractPyTorchNetwork):
+            def __init__(self):
+                super().__init__()
+                self.dense = torch.nn.Linear(1, 1)
+                self.relu = torch.nn.ReLU()
+
+            def forward(self, x):
+                return {"pred": self.relu(self.dense(x))}
+
+        return Model()
+
+    @staticmethod
+    def _setup_torchscript(*args):
+        import torch
+        from delira.models import AbstractTorchScriptNetwork
+
+        class Model(AbstractTorchScriptNetwork):
+            def __init__(self):
+                super().__init__()
+                self.dense = torch.nn.Linear(1, 1)
+                self.relu = torch.nn.ReLU()
+
+            @torch.jit.script_method
+            def forward(self, x):
+                return {"pred": self.relu(self.dense(x))}
+
+        return Model()
+
+    @staticmethod
+    def _setup_tfeager(*args):
+        import tensorflow as tf
+        from delira.models import AbstractTfEagerNetwork
+
+        class Model(AbstractTfEagerNetwork):
+            def __init__(self):
+                super().__init__()
+
+                self.dense = tf.keras.layers.Dense(1, activation="relu")
+
+            def call(self, x: tf.Tensor):
+                return {"pred": self.dense(x)}
+
+        return Model()
+
+    @staticmethod
+    def _setup_tfgraph(*args):
+        import tensorflow as tf
+        from delira.models import AbstractTfGraphNetwork
+        from delira.training.backends.tf_graph.utils import \
+            initialize_uninitialized
+
+        class Model(AbstractTfGraphNetwork):
+            def __init__(self):
+                super().__init__()
+                self.dense = tf.keras.layers.Dense(1, activation="relu")
+
+                data = tf.placeholder(shape=[None, 1],
+                                        dtype=tf.float32)
+
+                labels = tf.placeholder_with_default(
+                    tf.zeros([tf.shape(data)[0], 1]), shape=[None, 1])
+
+                preds_train = self.dense(data)
+                preds_eval = self.dense(data)
+
+                self.inputs["data"] = data
+                self.inputs["labels"] = labels
+                self.outputs_train["pred"] = preds_train
+                self.outputs_eval["pred"] = preds_eval
+
+        model = Model()
+        initialize_uninitialized(model._sess)
+        return model
+
+    @staticmethod
+    def _setup_chainer(*args):
+        import chainer
+        from delira.models import AbstractChainerNetwork
+
+        class Model(AbstractChainerNetwork):
+            def __init__(self):
+                super().__init__()
+
+                with self.init_scope():
+                    self.dense = chainer.links.Linear(1, 1)
+
+            def forward(self, x):
+                return {
+                    "pred":
+                        chainer.functions.relu(
+                            self.dense(x))
+                }
+
+        return Model()
+
+    @staticmethod
+    def _setup_sklearn(*args):
+
+        from delira.models import SklearnEstimator
+        from sklearn.neural_network import MLPRegressor
+
+        class Model(SklearnEstimator):
+            def __init__(self):
+                # prefit to enable prediction mode afterwards
+                module = MLPRegressor()
+                module.fit(*args)
+                super().__init__(module)
+
+            @staticmethod
+            def prepare_batch(batch: dict, input_device, output_device):
+                return batch
+
+        return Model()
+
+    def run_model_arg(self, device=None):
[…]
