### 2019-06-05 12ac946
initialize uninitialized members in TfExeriment.Test
