### 2019-06-05 d4d6a17
Initialize uninitialized variables in tf Experiment test
