### 2019-06-05 #1488 @whedon
[REVIEW]: Delira: A High-Level Framework for Deep Learning in Medical Image Analysis
**Submitting author:** @haarburger (<a href="http://orcid.org/0000-0002-2405-640X">Christoph Haarburger</a>)
**Repository:** <a href="https://github.com/justusschock/delira" target ="_blank">https://github.com/justusschock/delira</a>
**Version:** v0.3.2
