### 2019-06-05 98a347f
Add codecov after successful tests
