+++ .travis.yml
@@ -117,6 +117,11 @@ script:
           bash scripts/ci/run_style_checks.sh;
       fi
 
+after_success:
+    - if [[ "$TEST_TYPE" == "unittests" ]]; then
+          codecov;
+      fi
+
 before_deploy:
     - cd $TRAVIS_BUILD_DIR
 
