+++ delira/models/backends/torchscript/abstract_network.py
@@ -79,3 +79,97 @@ class AbstractTorchScriptNetwork(AbstractNetwork, torch.jit.ScriptModule):
                 torch.float)
 
         return return_dict
+
+    @staticmethod
+    def closure(model, data_dict: dict, optimizers: dict, losses={},
+                metrics={}, fold=0, **kwargs):
+        """
+        closure method to do a single backpropagation step
+
+        Parameters
+        ----------
+        model : :class:`AbstractTorchScriptNetwork`
+            trainable model
+        data_dict : dict
+            dictionary containing the data
+        optimizers : dict
+            dictionary of optimizers to optimize model's parameters
+        losses : dict
+            dict holding the losses to calculate errors
+            (gradients from different losses will be accumulated)
+        metrics : dict
+            dict holding the metrics to calculate
+        fold : int
+            Current Fold in Crossvalidation (default: 0)
+        **kwargs:
+            additional keyword arguments
+
+        Returns
+        -------
+        dict
+            Metric values (with same keys as input dict metrics)
+        dict
+            Loss values (with same keys as input dict losses)
+        list
+            Arbitrary number of predictions as torch.Tensor
+
+        Raises
+        ------
+        AssertionError
+            if optimizers or losses are empty or the optimizers are not
+            specified
+        """
+
+        assert (optimizers and losses) or not optimizers, \
+            "Criterion dict cannot be emtpy, if optimizers are passed"
+
+        loss_vals = {}
+        metric_vals = {}
+        total_loss = 0
+
+        # choose suitable context manager:
+        if optimizers:
+            context_man = torch.enable_grad
+
+        else:
+            context_man = torch.no_grad
+
+        with context_man():
+
+            # predict
+            inputs = data_dict.pop("data")
+            preds = model(inputs)
+
+            # calculate losses
+            for key, crit_fn in losses.items():
+                _loss_val = crit_fn(preds["pred"], data_dict["label"])
+                loss_vals[key] = _loss_val.item()
+                total_loss += _loss_val
+
+            # calculate metrics
+            with torch.no_grad():
+                for key, metric_fn in metrics.items():
+                    metric_vals[key] = metric_fn(
+                        preds["pred"], data_dict["label"]).item()
+
+        if optimizers:
+            optimizers['default'].zero_grad()
+            # apex does not yet support torchscript
+            total_loss.backward()
+            optimizers['default'].step()
+
+        else:
+
+            # add prefix "val" in validation mode
+            eval_loss_vals, eval_metrics_vals = {}, {}
+            for key in loss_vals.keys():
+                eval_loss_vals["val_" + str(key)] = loss_vals[key]
+
+            for key in metric_vals:
+                eval_metrics_vals["val_" + str(key)] = metric_vals[key]
+
+            loss_vals = eval_loss_vals
+            metric_vals = eval_metrics_vals
+
+        return metric_vals, loss_vals, {k: v.detach()
+                                        for k, v in preds.items()}
