### 2019-06-06 #1488 @PingjunChen
@haarburger please add DOI to references if exist. Also, the current reference format looks not standard.
