### 2019-06-06 7e14905
Add backend tests and fix minor bugs
