+++ tests/training/test_experiments.py
@@ -1,761 +1,0 @@
-
-from delira import get_backends
-import unittest
-import typing
-
-import numpy as np
-from functools import partial
-from delira.training import Parameters
-
-from delira.data_loading import AbstractDataset
-
-if "TF" in get_backends():
-    from delira.training.backends.tf_eager.utils import \
-        switch_tf_execution_mode
-
-
-class DummyDataset(AbstractDataset):
-    def __init__(self, length):
-        super().__init__(None, None)
-        self.length = length
-
-    def __getitem__(self, index):
-        return {"data": np.random.rand(32),
-                "label": np.random.randint(0, 1, 1)}
-
-    def __len__(self):
-        return self.length
-
-    def get_sample_from_index(self, index):
-        return self.__getitem__(index)
-
-
-if "CHAINER" in get_backends():
-    from delira.models import AbstractChainerNetwork
-    import chainer
-
-    # define this outside, because it has to be pickleable, which it won't be,
-    # wehn defined inside a function
-    class DummyNetworkChainer(AbstractChainerNetwork):
-        def __init__(self):
-            super().__init__()
-
-            with self.init_scope():
-                self.dense_1 = chainer.links.Linear(32, 64)
-                self.dense_2 = chainer.links.Linear(64, 1)
-
-        def forward(self, x):
-            return {
-                "pred":
-                    self.dense_2(chainer.functions.relu(
-                        self.dense_1(x)))
-            }
-
-
-class ExperimentTest(unittest.TestCase):
-
-    def setUp(self) -> None:
-
-        test_cases = {
-            "torch": [],
-            "torchscript": [],
-            "tf": [],
-            "tf_eager": [],
-            "chainer": [],
-            "sklearn": []
-        }
-
-        from sklearn.metrics import mean_absolute_error
-
-        # setup torch testcases
-        if "TORCH" in get_backends() and self._testMethodName.endswith("torch"):
-            import torch
-            from delira.models import AbstractPyTorchNetwork
-            from delira.training.callbacks import \
-                ReduceLROnPlateauCallbackPyTorch
-
-            class DummyNetworkTorch(AbstractPyTorchNetwork):
-
-                def __init__(self):
-                    super().__init__()
-                    self.module = self._build_model(32, 1)
-
-                def forward(self, x):
-                    return {"pred": self.module(x)}
-
-                @staticmethod
-                def _build_model(in_channels, n_outputs):
-                    return torch.nn.Sequential(
-                        torch.nn.Linear(in_channels, 64),
-                        torch.nn.ReLU(),
-                        torch.nn.Linear(64, n_outputs)
-                    )
-
-                @staticmethod
-                def prepare_batch(batch_dict, input_device, output_device):
-                    return {"data": torch.from_numpy(batch_dict["data"]
-                                                     ).to(input_device,
-                                                          torch.float),
-                            "label": torch.from_numpy(batch_dict["label"]
-                                                      ).to(output_device,
-                                                           torch.float)}
-
-            test_cases["torch"].append((
-                Parameters(fixed_params={
-                    "model": {},
-                    "training": {
-                        "losses": {"CE": torch.nn.BCEWithLogitsLoss()},
-                        "optimizer_cls": torch.optim.Adam,
-                        "optimizer_params": {"lr": 1e-3},
-                        "num_epochs": 2,
-                        "val_metrics": {"mae": mean_absolute_error},
-                        "lr_sched_cls": ReduceLROnPlateauCallbackPyTorch,
-                        "lr_sched_params": {"mode": "min"}
-                    }
-                }
-                ),
-                500,
-                50,
-                "mae",
-                "lowest",
-                DummyNetworkTorch))
-
[…]
