+++ tests/training/backends/test_torch.py
@@ -26,7 +26,7 @@ if "TORCH" in get_backends():
             }
 
 
-class TestChainerBackend(
+class TestTorchBackend(
     create_experiment_test_template_for_backend("TORCH")
 ):
     def setUp(self) -> None:
