+++ delira/models/backends/__init__.py
@@ -1,15 +1,15 @@
-from delira import get_backends
+from delira import get_backends as _get_backends
 
-if "CHAINER" in get_backends():
+if "CHAINER" in _get_backends():
     from .chainer import *
 
-if "SKLEARN" in get_backends():
+if "SKLEARN" in _get_backends():
     from .sklearn import *
 
-if "TF" in get_backends():
+if "TF" in _get_backends():
     from .tf_eager import *
     from .tf_graph import *
 
-if "TORCH" in get_backends():
+if "TORCH" in _get_backends():
     from .torch import *
     from .torchscript import *
