### 2019-06-06 5d217c1
Update unittests
