+++ delira/training/backends/tf_graph/trainer.py
@@ -3,7 +3,7 @@ from ..tf_eager.utils import create_optims_default, switch_tf_execution_mode
 from delira.training.utils import convert_to_numpy_identity as convert_to_numpy
 from delira.training.base_trainer import BaseNetworkTrainer
 from delira.io.tf import load_checkpoint, save_checkpoint
-from delira.models import AbstractTfGraphNetwork
+from delira.models.backends.tf_graph import AbstractTfGraphNetwork
 import os
 import logging
 
