+++ delira/models/backends/torch/__init__.py
@@ -1,3 +1,5 @@
-from .abstract_network import AbstractPyTorchNetwork
-from .data_parallel import DataParallelPyTorchNetwork
-from .utils import scale_loss
+from delira.models.backends.torch.abstract_network import \
+    AbstractPyTorchNetwork
+from delira.models.backends.torch.data_parallel import \
+    DataParallelPyTorchNetwork
+from delira.models.backends.torch.utils import scale_loss
