+++ delira/training/backends/sklearn/__init__.py
@@ -1,3 +1,4 @@
-from .trainer import SklearnEstimatorTrainer
-from .experiment import SklearnExperiment
-from .utils import create_optims_default
+from delira.training.backends.sklearn.trainer import SklearnEstimatorTrainer
+from delira.training.backends.sklearn.experiment import SklearnExperiment
+from delira.training.backends.sklearn.utils import create_optims_default \
+    as create_sklearn_optims_default
