+++ delira/io/tf.py
@@ -1,4 +1,4 @@
-from ..models.backends.tf_eager import AbstractTfEagerNetwork
+from delira.models.backends.tf_eager import AbstractTfEagerNetwork
 import typing
 import logging
 
