+++ tests/data_loading/test_dataset.py
@@ -9,23 +9,24 @@ from delira.data_loading.load_utils import norm_zero_mean_unit_std
 
 class DataSubsetConcatTest(unittest.TestCase):
 
+    @staticmethod
+    def load_dummy_sample(path, label_load_fct):
+        """
+        Returns dummy data, independent of path or label_load_fct
+        Parameters
+        ----------
+        path
+        label_load_fct
+        Returns
+        -------
+        : dict
+            dict with data and label
+        """
+
+        return {'data': np.random.rand(1, 256, 256),
+                'label': np.random.randint(2)}
+
     def test_data_subset_concat(self):
-
-        def load_dummy_sample(path, label_load_fct):
-            """
-            Returns dummy data, independent of path or label_load_fct
-            Parameters
-            ----------
-            path
-            label_load_fct
-            Returns
-            -------
-            : dict
-                dict with data and label
-            """
-
-            return {'data': np.random.rand(1, 256, 256),
-                    'label': np.random.randint(2)}
 
         class DummyCacheDataset(BaseCacheDataset):
             def __init__(self, num: int, label_load_fct, *args, **kwargs):
@@ -50,9 +51,9 @@ class DataSubsetConcatTest(unittest.TestCase):
                     data.append(self._load_fn(i, self.label_load_fct))
                 return data
 
-        dset_a = DummyCacheDataset(500, None, load_fn=load_dummy_sample,
+        dset_a = DummyCacheDataset(500, None, load_fn=self.load_dummy_sample,
                                    img_extensions=[], gt_extensions=[])
-        dset_b = DummyCacheDataset(700, None, load_fn=load_dummy_sample,
+        dset_b = DummyCacheDataset(700, None, load_fn=self.load_dummy_sample,
                                    img_extensions=[], gt_extensions=[])
 
         # test concatenating
@@ -77,130 +78,129 @@ class DataSubsetConcatTest(unittest.TestCase):
         # check if entries are valid
         self.assertTrue(sliced_concat_set[0])
 
+    def test_cache_dataset(self):
+        def load_mul_sample(path):
+            """
+            Return a list of random samples
+            Parameters
+            ----------
+            path
 
-def test_cache_dataset():
-    def load_mul_sample(path):
-        """
-        Return a list of random samples
-        Parameters
-        ----------
-        path
+            Returns
+            -------
+            list
+                list of samples
+            """
+            return [self.load_dummy_sample(path, None)] * 4
 
-        Returns
-        -------
-        list
-            list of samples
-        """
-        return [load_dummy_sample(None)] * 4
+        # test normal cache dataset
+        paths = list(range(10))
+        dataset = BaseCacheDataset(paths, self.load_dummy_sample,
+                                   label_load_fct=None)
+        assert len(dataset) == 10
+        try:
+            a = dataset[0]
+            a = dataset[5]
+            a = dataset[9]
+        except BaseException:
+            raise AssertionError('Dataset access failed.')
 
-    # test normal cache dataset
-    paths = list(range(10))
-    dataset = BaseCacheDataset(paths, load_dummy_sample)
-    assert len(dataset) == 10
-    try:
-        a = dataset[0]
-        a = dataset[5]
-        a = dataset[9]
-    except BaseException:
-        raise AssertionError('Dataset access failed.')
+        try:
+            j = 0
+            for i in dataset:
+                assert 'data' in i
+                assert 'label' in i
+                j += 1
+            assert j == len(dataset)
+        except BaseException:
+            raise AssertionError('Dataset iteration failed.')
 
-    try:
[…]
