+++ delira/models/backends/chainer/__init__.py
@@ -1,4 +1,10 @@
-from .abstract_network import AbstractChainerNetwork
-from .data_parallel import DataParallelChainerNetwork, DataParallelChainerOptimizer, \
-    ParallelOptimizerUpdateModelParameters, \
+from delira.models.backends.chainer.abstract_network import \
+    AbstractChainerNetwork
+from delira.models.backends.chainer.data_parallel import \
+    DataParallelChainerNetwork
+from delira.models.backends.chainer.data_parallel import \
+    DataParallelChainerOptimizer
+from delira.models.backends.chainer.data_parallel import \
+    ParallelOptimizerUpdateModelParameters
+from delira.models.backends.chainer.data_parallel import \
     ParallelOptimizerCumulateGradientsHook
