### 2019-06-06 aa836f9
Add additional guards for imports
