+++ delira/models/backends/tf_eager/__init__.py
@@ -1,4 +1,7 @@
-from delira.models.backends.tf_eager.abstract_network import \
-    AbstractTfEagerNetwork
-from delira.models.backends.tf_eager.data_parallel import \
-    DataParallelTfEagerNetwork
+from delira import get_backends as _get_backends
+
+if "TF" in _get_backends():
+    from delira.models.backends.tf_eager.abstract_network import \
+        AbstractTfEagerNetwork
+    from delira.models.backends.tf_eager.data_parallel import \
+        DataParallelTfEagerNetwork
