+++ delira/models/backends/torchscript/__init__.py
@@ -1,1 +1,4 @@
-from .abstract_network import AbstractTorchScriptNetwork
+from delira import get_backends as _get_backends
+
+if "TORCH" in _get_backends():
+    from .abstract_network import AbstractTorchScriptNetwork
