+++ delira/training/backends/tf_graph/__init__.py
@@ -1,3 +1,7 @@
-from delira.training.backends.tf_graph.experiment import TfGraphExperiment
-from delira.training.backends.tf_graph.trainer import TfGraphNetworkTrainer
-from delira.training.backends.tf_graph.utils import initialize_uninitialized
+from delira import get_backends as _get_backends
+
+if "TF" in _get_backends():
+    from delira.training.backends.tf_graph.experiment import TfGraphExperiment
+    from delira.training.backends.tf_graph.trainer import TfGraphNetworkTrainer
+    from delira.training.backends.tf_graph.utils import \
+        initialize_uninitialized
