+++ delira/training/backends/chainer/__init__.py
@@ -1,6 +1,9 @@
-from delira.training.backends.chainer.trainer import ChainerNetworkTrainer
-from delira.training.backends.chainer.experiment import ChainerExperiment
-from delira.training.backends.chainer.utils import convert_to_numpy \
-    as convert_chainer_to_numpy
-from delira.training.backends.chainer.utils import create_optims_default \
-    as create_chainer_optims_default
+from delira import get_backends as _get_backends
+
+if "CHAINER" in _get_backends():
+    from delira.training.backends.chainer.trainer import ChainerNetworkTrainer
+    from delira.training.backends.chainer.experiment import ChainerExperiment
+    from delira.training.backends.chainer.utils import convert_to_numpy \
+        as convert_chainer_to_numpy
+    from delira.training.backends.chainer.utils import create_optims_default \
+        as create_chainer_optims_default
