+++ delira/training/backends/torch/__init__.py
@@ -1,6 +1,9 @@
-from delira.training.backends.torch.trainer import PyTorchNetworkTrainer
-from delira.training.backends.torch.experiment import PyTorchExperiment
-from delira.training.backends.torch.utils import create_optims_default \
-    as create_pytorch_optims_default
-from delira.training.backends.torch.utils import convert_to_numpy \
-    as convert_torch_to_numpy
+from delira import get_backends as _get_backends
+
+if "TORCH" in _get_backends():
+    from delira.training.backends.torch.trainer import PyTorchNetworkTrainer
+    from delira.training.backends.torch.experiment import PyTorchExperiment
+    from delira.training.backends.torch.utils import create_optims_default \
+        as create_pytorch_optims_default
+    from delira.training.backends.torch.utils import convert_to_numpy \
+        as convert_torch_to_numpy
