+++ delira/training/backends/sklearn/__init__.py
@@ -1,4 +1,8 @@
-from delira.training.backends.sklearn.trainer import SklearnEstimatorTrainer
-from delira.training.backends.sklearn.experiment import SklearnExperiment
-from delira.training.backends.sklearn.utils import create_optims_default \
-    as create_sklearn_optims_default
+from delira import get_backends as _get_backends
+
+if "SKLEARN" in _get_backends():
+    from delira.training.backends.sklearn.trainer import \
+        SklearnEstimatorTrainer
+    from delira.training.backends.sklearn.experiment import SklearnExperiment
+    from delira.training.backends.sklearn.utils import create_optims_default \
+        as create_sklearn_optims_default
