+++ tests/training/test_metrics.py
@@ -35,6 +35,5 @@ class TestMetrics(unittest.TestCase):
         self.assertEqual(score_auc, 0.5)
 
 
-
 if __name__ == '__main__':
     unittest.main()
