### 2019-06-06 dea575d
Fix metric test
