### 2019-06-06 7e41917
add switch between eager and graph execution for tf io tests
