### 2019-06-06 62982ae
minor test fixes
