### 2019-06-06 94282fe
Add minor bugfixes tf
