### 2019-06-06 1c2aaf4
add dtype conversions for tf
