+++ tests/io/test_tf.py
@@ -91,8 +91,8 @@ class IoTfTest(unittest.TestCase):
                 return self.net(inputs)
 
         net = DummyNetwork((32,), 1)
-        input_tensor = np.random.rand(1, 32)
-        result_pre_save = net(input_tensor)
+        input_tensor = np.random.rand(1, 32).astype(np.float32)
+        result_pre_save = net()
         save_checkpoint_eager("./model_eager", model=net)
 
         loaded_state = load_checkpoint_eager("./model_eager", model=net)
