### 2019-06-06 7e1711d
Reset TFs default graph
