+++ delira/training/backends/tf_eager/experiment.py
@@ -260,3 +260,36 @@ class TfEagerExperiment(BaseExperiment):
                             metrics=metrics, metric_keys=metric_keys,
                             verbose=verbose, prepare_batch=prepare_batch,
                             convert_fn=convert_fn, **kwargs)
+
+    def setup(self, params, training=True, **kwargs):
+        """
+        Defines the setup behavior (model, trainer etc.) for training and
+        testing case
+
+        Parameters
+        ----------
+        params : :class:`Parameters`
+            the parameters to use for setup
+        training : bool
+            whether to setup for training case or for testing case
+        **kwargs :
+            additional keyword arguments
+
+        Returns
+        -------
+        :class:`BaseNetworkTrainer`
+            the created trainer (if ``training=True``)
+        :class:`Predictor`
+            the created predictor (if ``training=False``)
+
+        See Also
+        --------
+
+        * :meth:`BaseExperiment._setup_training` for training setup
+
+        * :meth:`BaseExperiment._setup_test` for test setup
+
+        """
+        tf.reset_default_graph()
+        return super().setup(params=params, training=training,
+                             **kwargs)
