### 2019-06-06 #1488 @haarburger
@PingjunChen, unfortunately none of the references has a DOI. Please let me know if you have further suggestions on how to improve the reference format.
