### 2019-06-06 #1488 @PingjunChen
@haarburger Thanks, now the reference format should be fine.

I have some difficulty to understand the installation on the backend. Take tensorflow as an example,  does ```pip install delira[tf]``` mean ```pip install delira tf```? But in PyPI, `tf` is not `tensorflow` . Please give more explicit installation instructions.

Because of the complexity of this framework, I need more time to check the functionalities and tests.
