### 2019-06-07 ecc1286
convert to lookupconfig
