+++ delira/utils/config.py
@@ -64,10 +64,7 @@ class LookupConfig(Config):
 
         """
 
-        if type(value) == dict or type(value) == Config:
-            update_val = LookupConfig()
-            update_val.update(value, deep=False)
-        else:
-            update_val = value
+        if isinstance(value, dict) and not isinstance(value, LookupConfig):
+            value = LookupConfig(config=value)
 
-        return super().__setattr__(key, update_val)
+        return super().__setattr__(key, value)
