### 2019-06-07 567a7ea
Fix datamanager test
