### 2019-06-07 a412ee3
Fix minor test bugs
