### 2019-06-07 #1488 @haarburger
@PingjunChen the pip command is supposed to be used exactly as stated in our readme. Maybe [this](https://stackoverflow.com/questions/46775346/what-do-square-brackets-mean-in-pip-install) explanation about the syntax is helpful. It means that we want to install the `tensorflow`/`pytorch` backend as an extra.
