### 2019-06-07 49e02cf
Try to fix tf failure by explicitly deleting it after each test
