+++ tests/training/backends/test_tf_eager.py
@@ -1,4 +1,5 @@
 import unittest
+import gc
 from delira import get_backends
 from delira.training import Parameters
 from sklearn.metrics import mean_absolute_error
@@ -74,6 +75,10 @@ class TestTfEagerBackend(
         self._experiment_cls = experiment_cls
 
         super().setUp()
+
+    def tearDown(self):
+        del tf
+        gc.collect()
 
 
 if __name__ == "__main__":
