+++ tests/training/backends/test_tf_eager.py
@@ -77,7 +77,11 @@ class TestTfEagerBackend(
         super().setUp()
 
     def tearDown(self):
-        del tf
+        try:
+            del tf
+            del tensorflow
+        except (UnboundLocalError, NameError):
+            pass
         gc.collect()
 
 
