### 2019-06-07 #1488 @PingjunChen
@haarburger The link you provide helps me understand the usage of the square bracket. Although for deep learning people,  `tf` defaults to `tensorflow`, and you also give `tensorflow` link to `tf`. Nevertheless, I still suggest changing `tf` to `tensorflow`, at least in the `Installation` part. The main reason is to avoid ambiguity, as [`tf`](https://pypi.org/project/tf/) is already used by https://pypi.org/project/tf in PyPI.
