### 2019-06-09 #1488 @acolum
I'm finished with my review and am recommending this for publication. @haarburger I just wanted to commend you and your team for a job well done writing and documenting this package. This is one of the best JOSS submissions I've reviewed.
