### 2019-06-09 #1488 @PingjunChen
@mbobra Just a reminder that we have finished the review. Overall, this package has rich features for deep learning model training, data preparation, visualization, etc. In addition, the documentation is well-written with iPython notebook examples for classification, segmentation, and GAN applications. I recommend accepting for publication.
