### 2019-06-10 justusschock/delira#126 @haarburger
Rename installation instructions for tensorflow backend
One of our reviewers at JOSS [pointed out](https://github.com/openjournals/joss-reviews/issues/1488#issuecomment-499916993) that installing `delira` with `tensorflow` backend with `pip install delira[tf]` is ambiguous and confusing because there is already `tf` package registered at PyPI.

We should change that to `pip install delira[tensorflow]`.

Does any of you disagree @justusschock @mibaumgartner @ORippler ?
