### 2019-06-10 #1488 @arfon
@haarburger - At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? Please make sure that the title of the Zenodo archive and authors match the authors of the paper here.

Once you've done this, I can then move forward with accepting the submission.
