### 2019-06-10 #1488 @arfon
> I think the software also needs a new tagged release with the latest changes?

Oh yes. Please do that too when making the Zenodo archive!
