### 2019-06-11 0b80145
Fix predictor caching behavior (disabled during training; enabled during testing)
