+++ delira/training/predictor.py
@@ -139,10 +139,10 @@ class Predictor(object):
         )[1]
 
     def predict_data_mgr(self, datamgr, batchsize=None, metrics={},
-                         metric_keys=None, verbose=False, lazy_gen=False,
-                         **kwargs):
+                         metric_keys=None, verbose=False, **kwargs):
         """
         Defines a routine to predict data obtained from a batchgenerator
+        without explicitly caching anything
 
         Parameters
         ----------
@@ -158,10 +158,6 @@ class Predictor(object):
             the ``batch_dict`` items to use for metric calculation
         verbose : bool
             whether to show a progress-bar or not, default: False
-        lazy_gen : bool
-            if True: Yields results instead of returning them; should be
-            specified if predicting on a low-memory device or when results
-            should be saved immediately
         kwargs :
             keyword arguments passed to :func:`prepare_batch_fn`
 
@@ -169,16 +165,8 @@ class Predictor(object):
         ------
         dict
             a dictionary containing all predictions of the current batch
-            if ``lazy_gen`` is True
         dict
             a dictionary containing all metrics of the current batch
-            if ``lazy_gen`` is True
-        dict
-            a dictionary containing all predictions;
-            if ``lazy_gen`` is False
-        dict
-            a dictionary containing all validation metrics (maybe empty);
-            if ``lazy_gen`` is False
 
         """
 
@@ -192,9 +180,6 @@ class Predictor(object):
         datamgr.n_process_augmentation = 1
 
         batchgen = datamgr.get_batchgen()
-
-        if not lazy_gen:
-            predictions_all, metric_vals = [], {k: [] for k in metrics.keys()}
 
         n_batches = batchgen.num_batches * batchgen.num_processes
 
@@ -245,13 +230,7 @@ class Predictor(object):
                                                  metrics=metrics,
                                                  metric_keys=metric_keys)
 
-                if lazy_gen:
-                    yield preds, _metric_vals
-                else:
-                    for k, v in _metric_vals.items():
-                        metric_vals[k].append(v)
-
-                    predictions_all.append(preds)
+                yield preds, _metric_vals
 
                 batch_list = []
 
@@ -259,97 +238,242 @@ class Predictor(object):
         datamgr.batch_size = orig_batch_size
         datamgr.n_process_augmentation = orig_num_aug_processes
 
-        if lazy_gen:
-            # triggers stopiteration
-            return
+        return
 
-        # convert predictions from list of dicts to dict of lists
-        new_predictions_all = {}
+    def predict_data_mgr_cache_metrics_only(self, datamgr, batchsize=None,
+                                            metrics={}, metric_keys=None,
+                                            verbose=False, **kwargs):
+        """
+        Defines a routine to predict data obtained from a batchgenerator and
+        caches the metrics
 
-        def __convert_dict(old_dict, new_dict):
-            """
-            Function to recursively convert dicts
+        Parameters
+        ----------
+        datamgr : :class:`BaseDataManager`
+            Manager producing a generator holding the batches
+        batchsize : int
+            Artificial batchsize (sampling will be done with batchsize
+            1 and sampled data will be stacked to match the artificial
+            batchsize)(default: None)
+        metrics : dict
+            the metrics to calculate
+        metric_keys : dict
+            the ``batch_dict`` items to use for metric calculation
+        verbose : bool
+            whether to show a progress-bar or not, default: False
+        kwargs :
+            keyword arguments passed to :func:`prepare_batch_fn`
 
-            Parameters
-            ----------
-            old_dict : dict
-                the old nested dict
-            new_dict : dict
-                the new nested dict
+        Yields
+        ------
+        dict
+            a dictionary containing all validation metrics (maybe empty)
 
-            Returns
-            -------
-            dict
-                the updated new nested dict
[…]
