### 2019-06-11 4727261
remove second variable in basetrainer; fixes unpacking error
