### 2019-06-11 47b5d80
Fix num batches behavior based on debug mode
