+++ delira/data_loading/sampler/sequential_sampler.py
@@ -20,10 +20,9 @@ class SequentialSampler(AbstractSampler):
             data index and the value at a certain index indicates the
             corresponding class
         """
-        super().__init__()
+        super().__init__(indices)
 
         self._indices = list(range(len(indices)))
-        self._global_index = 0
 
     def _get_indices(self, n_indices):
         """
@@ -43,23 +42,12 @@ class SequentialSampler(AbstractSampler):
         list
             list of sampled indices
         """
-        if self._global_index >= len(self._indices):
-            self._global_index = 0
-            raise StopIteration
+        n_indices = self._check_batchsize(n_indices)
 
-        new_global_idx = self._global_index + n_indices
-
-        # If we reach end, make batch smaller
-        if new_global_idx >= len(self._indices):
-            new_global_idx = len(self._indices)
-
-        idxs = list(range(self._global_index, new_global_idx))
-        self._global_index = new_global_idx
-
-        return idxs
+        return range(self._global_index - n_indices, self._global_index)
 
     def __len__(self):
-        return len(self._indices)
+        return self._num_samples
 
 
 class PrevalenceSequentialSampler(AbstractSampler):
@@ -86,7 +74,7 @@ class PrevalenceSequentialSampler(AbstractSampler):
             if True: indices will be sampled in a sequential way per class and
             sampled indices will be shuffled
         """
-        super().__init__()
+        super().__init__(indices)
 
         _indices = {}
         _global_idxs = {}
@@ -107,12 +95,11 @@ class PrevalenceSequentialSampler(AbstractSampler):
             ordered_dict[k] = _indices[k]
             length += len(_indices[k])
 
-        self._length = length
+        self._num_samples = length
 
         self._indices = ordered_dict
         self._n_classes = len(_indices.keys())
         self._global_idxs = _global_idxs
-        self._global_total_idx = 0
 
         self._shuffle = shuffle_batch
 
@@ -140,11 +127,9 @@ class PrevalenceSequentialSampler(AbstractSampler):
         list
             list of sampled indices
         """
-        samples_per_class = int(n_indices / self._n_classes)
 
-        if self._global_total_idx >= len(self._indices):
-            self._global_total_idx = 0
-            raise StopIteration
+        n_indices = self._check_batchsize(n_indices)
+        samples_per_class = n_indices // self._n_classes
 
         _samples = []
 
@@ -172,15 +157,13 @@ class PrevalenceSequentialSampler(AbstractSampler):
             _samples += list(range(self._global_idxs[key], new_global_idx))
             self._global_idxs[key] = new_global_idx
 
-        self._global_total_idx += n_indices
-
         if self._shuffle:
             shuffle(_samples)
 
         return _samples
 
     def __len__(self):
-        return self._length
+        return self._num_samples
 
 
 class StoppingPrevalenceSequentialSampler(AbstractSampler):
@@ -242,6 +225,58 @@ class StoppingPrevalenceSequentialSampler(AbstractSampler):
         labels = [dataset[idx]['label'] for idx in indices]
         return cls(labels)
 
+    def _check_batchsize(self, n_indices):
+        """
+        Checks if batchsize is valid for all classes
+
+        Parameters
+        ----------
+        n_indices : int
+            the number of samples to return
+
+        Returns
+        -------
+        dict
+            number of samples per class to return
+
+        """
+        n_indices = super()._check_batchsize(n_indices)
+
+        samples_per_class = n_indices // self._n_classes
+        remaining = n_indices % self._n_classes
+
+        samples = {}
+
+        try:
+
+            # sample same number of sample for each class
+            for key, idx_list in self._indices.items():
+                if self._global_idxs[key] >= len(idx_list):
+                    raise StopIteration
+
+                # truncate if necessary
[…]
