### 2019-06-11 3a70c84
Rewrite Samplers
