+++ docs/getting_started.rst
@@ -14,10 +14,10 @@ The currently supported backends are:
   .. note::
     ``delira`` supports mixed-precision training via `apex <https://github.com/NVIDIA/apex>`_, but ``apex`` must be installed separately
     
-* `tf <https://tensorflow.org>`_ (very experimental): Suffix ``tf``
+* `tf <https://tensorflow.org>`_ (very experimental): Suffix ``tensorflow``
 
   .. note::
-    the ``tf`` backend is still very experimental and may be unstable.
+    the ``tensorflow`` backend is still very experimental and may be unstable.
 
 * None: No Suffix
 
@@ -32,16 +32,60 @@ The currently supported backends are:
 Installation
 ------------
 
-======== ============================= ============================================================================================= ======================================================================================================================
-Backend  Binary Installation           Source Installation                                                                           Notes
-======== ============================= ============================================================================================= ======================================================================================================================
-None     ``pip install delira``        ``pip install git+https://github.com/justusschock/delira.git``                                Training not possible if backend is not installed separately
-`torch`_ ``pip install delira[torch]`` ``git clone https://github.com/justusschock/delira.git && cd delira && pip install .[torch]`` ``delira`` with ``torch`` backend supports mixed-precision training via `NVIDIA/apex`_ (must be installed separately).
-`tf`_               --                 ``git clone https://github.com/justusschock/delira.git && cd delira && pip install .[tf]``    The tensorflow backend is still very experimental and lacks some `features`_
-Full     ``pip install delira[full]``  ``git clone https://github.com/justusschock/delira.git && cd delira && pip install .[full]``  All backends are getting installed.
-======== ============================= ============================================================================================= ======================================================================================================================
++---------+-----+--------------------+---------------------------------+
+| Backend | Bin | Source             | Notes                           |
+|         | ary | Installation       |                                 |
+|         | Ins |                    |                                 |
+|         | tal |                    |                                 |
+|         | lat |                    |                                 |
+|         | ion |                    |                                 |
++=========+=====+====================+=================================+
+| None    | ``p | ``pip install git+ | Training not possible if        |
+|         | ip  | https://github.com | backend is not installed        |
+|         | ins | /justusschock/deli | separately                      |
+|         | tal | ra.git``           |                                 |
+|         | l d |                    |                                 |
+|         | eli |                    |                                 |
+|         | ra` |                    |                                 |
+|         | `   |                    |                                 |
++---------+-----+--------------------+---------------------------------+
+| `torch` | ``p | ``git clone https: | ``delira`` with ``torch``       |
+| _       | ip  | //github.com/justu | backend supports                |
+|         | ins | sschock/delira.git | mixed-precision training via    |
+|         | tal |  && cd delira && p | `NVIDIA/apex`_ (must be         |
+|         | l d | ip install .[torch | installed separately).          |
+|         | eli | ]``                |                                 |
+|         | ra[ |                    |                                 |
+|         | tor |                    |                                 |
+|         | ch] |                    |                                 |
+|         | ``  |                    |                                 |
++---------+-----+--------------------+---------------------------------+
+| `tensor | ``p | ``git clone https: | the ``tensorflow`` backend is   |
+| flow`_  | ip  | //github.com/justu | still very experimental and     |
+|         | ins | sschock/delira.git | lacks some `features`_          |
+|         | tal |  && cd delira && p |                                 |
+|         | l d | ip install .[tenso |                                 |
+|         | eli | rflow]``           |                                 |
+|         | ra[ |                    |                                 |
+|         | ten |                    |                                 |
+|         | sor |                    |                                 |
+|         | flo |                    |                                 |
[…]
