### 2019-06-11 justusschock/delira#136 @justusschock
New Release
According to https://github.com/openjournals/joss-reviews/issues/1488 we need to create a new release, before we make an archive of our repo.
I'd like to have #132 #131 #130 #129 #128 #127 #126 as part of this release.

Most of these changes are either already implemented and awaiting review or require only minor changes. I'm aiming for a release date tomorrow evening (at latest).

Are you okay with this? Did I miss something which must be included? Do you want to postpone the release/have it earlier?

cc @haarburger @mibaumgartner @ORippler
