### 2019-06-11 898a2f0
fix samplers
