### 2019-06-11 #1488 @haarburger
We'd like to add a few bugfixes and minor improvements to the new tag. I'll let you know when these changes are merged and tagged and the Zenodo archive is created.
