### 2019-06-11 eb53955
best_val_score fix, interchangeable predcitor cls, new metric_keys docstring
