### 2019-06-12 8d7dd47
Update requirements.txt
Remove pinned scipy requirement (awating new trixi release)
