### 2019-06-12 justusschock/delira#136 @justusschock
We should be ready for release now. I'll draft some release notes here and tag the release later today.
