+++ delira/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.3.3'
+__version__ = '0.4.0'
 
 import json
 import os
