### 2019-06-12 justusschock/delira#136 @justusschock
So here is the promised draft. I hope I did not forget anything we changed.

---

# Release v0.4.0

## New Features
* **Slack**: We now have a community chat on [slack](https://join.slack.com/t/deliradev/shared_invite/enQtNjI1MjA4MjQzMzQ2LTUzNTQ0MjQyNjJjNzgyODczY2Y1YjYxNjA3ZmQ0MGFhODhkYzQ4M2RjMGM1YWM3YWU5MDM0ZjdiNTQ4MmQ0ZDk) [[commit](e3307db4d450f9c37d45d9c838819021418e3b1a)]
* **Paper**: We wrote and submitted a Paper to [JOSS](https://joss.theoj.org/) [[commit](9af1915df077fdc52901b6a1696c7f6d0efd77a0)]
* **Dataset Integration Guide**: New tutorial to switch from old dataset API to new dataset API [[commit](42b55c388e522b75451e86055ae943acfdafcc7a)]
* **Predictor**: Introduce new `Predictor` class to predict from network without having to initialize Trainer with dummy values [[commit](69080a87fa2255d6e8da3564bd11b157e7e00458)]
* **PEP-8**: Complete Codebase is now PEP-8 compatible and CI/CD now checks for it [[commit](5db9a6533129f4aebb60dc8a0587a7616313567c)]
* **Debug Mode**: New `debug mode` disables all multiprocessing (currently multiprocessing is only used during data loading); Might slow down the program but increases debug capabilities [[commit](b5713ddbe85bb356a4c7f97d5587b3c69750d063)]

## Non-Breaking Changes
* **Unify Trainer**: Merging Code from Backend Trainers (`TfNetworkTrainer` and `PyTorchNetworkTrainer`) into `BaseNetworkTrainer`; Will increase maintainability and ease of adding new backends [[commit](69080a87fa2255d6e8da3564bd11b157e7e00458)]
* **Unify Experiment**: Merging Code from Backend Experiments to `BaseExperiment` [[commit](69080a87fa2255d6e8da3564bd11b157e7e00458)]
* **Sampler StopIteration**: Refactor samplers to use a common function to check if `StopIteration` must be raise (where possible) [[commit](9d25a68516757cd39482939cd599bb267727a4b6)]

## Bugfixes
* **TF Resnet18**: ResNet18 in TF now matches the official paper [[commit](9a133c0bdfe148f861f013b756be70b76d480c4a)]
* **Parameters**: Correct Permutations in `Parameters` class [[commit](60fa85d7dcc07928c03b1db3ffdec5ab6428d0b8)]
* **Infinite Recursion**:  Infinite recursion in `LookUpConfig` due to wrong typechecks is now fixed [[commit](ecc1286507942ecc6625c06287cc6e07c8acf0ce)]
* **DataLoading `num_batches`**: The number of batches is now calculated without any influence of `multiprocessing` [[commit](9d25a68516757cd39482939cd599bb267727a4b6)]

## Breaking Changes
* **Return dicts**: Models must now return `dict`s during prediction and training [[commit](6f023a8365dd2007a3d22862b7b4fb95652d77b5)]
* **Kfold**: The old `kfold`- function has now merged with the new `kfold`, `stratified_kfold` and `stratified_kfold_predict` to a new `kfold` function supporting all these behaviors. [[commit](69080a87fa2255d6e8da3564bd11b157e7e00458)]
* **Rename tf Backend**: The `tf` backend has been renamed to `tensorflow` during installation. The new command to install `delira` with TensorFlow backend is now: `pip install delira[tensorflow]` [[commit](74e0d6dd19d805d6ab84d719cd49ec322c3f9f36)]
