### 2019-06-12 70f6b9e
Add new requirement files to MANIFEST.in
