+++ delira/io/torch.py
@@ -1,168 +1,167 @@
 import logging
 import os
 from collections import OrderedDict
-from delira import get_backends
 
 logger = logging.getLogger(__name__)
 
-if "TORCH" in get_backends():
+import torch
+from delira.models.backends.torch import AbstractPyTorchNetwork
+from delira.models.backends.torchscript import AbstractTorchScriptNetwork
 
-    import torch
-    from delira.models.backends.torch import AbstractPyTorchNetwork
-    from delira.models.backends.torchscript import AbstractTorchScriptNetwork
 
-    def save_checkpoint_torch(file: str, model=None, optimizers={},
-                              epoch=None, **kwargs):
-        """
-        Save checkpoint
+def save_checkpoint_torch(file: str, model=None, optimizers={},
+                          epoch=None, **kwargs):
+    """
+    Save checkpoint
 
-        Parameters
-        ----------
-        file : str
-            filepath the model should be saved to
-        model : AbstractNetwork or None
-            the model which should be saved
-            if None: empty dict will be saved as state dict
-        optimizers : dict
-            dictionary containing all optimizers
-        epoch : int
-            current epoch (will also be pickled)
+    Parameters
+    ----------
+    file : str
+        filepath the model should be saved to
+    model : AbstractNetwork or None
+        the model which should be saved
+        if None: empty dict will be saved as state dict
+    optimizers : dict
+        dictionary containing all optimizers
+    epoch : int
+        current epoch (will also be pickled)
 
-        """
-        if isinstance(model, torch.nn.DataParallel):
-            _model = model.module
-        else:
-            _model = model
+    """
+    if isinstance(model, torch.nn.DataParallel):
+        _model = model.module
+    else:
+        _model = model
 
-        if isinstance(_model, (AbstractPyTorchNetwork,
-                               AbstractTorchScriptNetwork)):
-            model_state = _model.state_dict()
-        else:
-            model_state = {}
-            logger.debug("Saving checkpoint without Model")
+    if isinstance(_model, (AbstractPyTorchNetwork,
+                           AbstractTorchScriptNetwork)):
+        model_state = _model.state_dict()
+    else:
+        model_state = {}
+        logger.debug("Saving checkpoint without Model")
 
-        optim_state = OrderedDict()
-        for key, val in optimizers.items():
-            if isinstance(val, torch.optim.Optimizer):
-                optim_state[key] = val.state_dict()
+    optim_state = OrderedDict()
+    for key, val in optimizers.items():
+        if isinstance(val, torch.optim.Optimizer):
+            optim_state[key] = val.state_dict()
 
-        if not optim_state:
-            logger.debug("Saving checkpoint without Optimizer")
+    if not optim_state:
+        logger.debug("Saving checkpoint without Optimizer")
 
-        if epoch is None:
-            epoch = 0
+    if epoch is None:
+        epoch = 0
 
-        state = {"optimizer": optim_state,
-                 "model": model_state,
-                 "epoch": epoch}
+    state = {"optimizer": optim_state,
+             "model": model_state,
+             "epoch": epoch}
 
-        torch.save(state, file, **kwargs)
+    torch.save(state, file, **kwargs)
 
-    def load_checkpoint_torch(file, **kwargs):
-        """
-        Loads a saved model
+def load_checkpoint_torch(file, **kwargs):
+    """
+    Loads a saved model
 
-        Parameters
-        ----------
-        file : str
-            filepath to a file containing a saved model
-        **kwargs:
-            Additional keyword arguments (passed to torch.load)
-            Especially "map_location" is important to change the device the
-            state_dict should be loaded to
+    Parameters
+    ----------
+    file : str
+        filepath to a file containing a saved model
[…]
