+++ requirements_extra_tf.txt
@@ -1,3 +1,3 @@
 tensorflow-gpu==1.13.1
 # use trixi-slim >= 0.1.2.1 for tensorboardX support
-trixi-slim>=0.1.2.1
+trixi-slim>0.1.2.1
