+++ requirements.txt
@@ -10,7 +10,7 @@ pylint
 tqdm
 visdom>=0.1.8.5
 pyyaml
-trixi-slim>0.1.2.0
+trixi-slim>0.1.2.1
 batchgenerators>=0.18.2,!=0.19.2
 psutil
 nested_lookup
