### 2019-06-12 justusschock/delira#136 @justusschock
Release is tagged [here](https://github.com/justusschock/delira/tree/v0.4.0)
