+++ delira/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.4.1'
+__version__ = '0.4.0'
 
 import json
 import os
