### 2019-06-12 cfa804c
Update .travis.yml
