### 2019-06-12 8d3de8c
Add basic single-threaded logging
