+++ tests/logging/test_single_threaded_logging.py
@@ -1,0 +1,515 @@
+from delira.logging import Logger, TensorboardBackend, make_logger
+
+import unittest
+
+try:
+    import torch
+except ImportError:
+    torch = None
+
+try:
+    import tensorflow as tf
+except ImportError:
+    tf = None
+
+try:
+    import onnx
+except ImportError:
+    onnx = None
+
+import numpy as np
+import os
+import gc
+
+
+class TestTensorboardLogging(unittest.TestCase):
+
+    def setUp(self) -> None:
+
+        self._npy_imgs = np.random.rand(2, 3, 24, 24)
+        self._boxes_npy = np.array([[5, 5, 10, 10], [4, 8, 5, 16]])
+        self._scalars = [{"1": 4, "2": 14, "3": 24},
+                         {"1": 5, "2": 15, "3": 25},
+                         {"1": 6, "2": 16, "3": 26}]
+
+        self._hist_vals = np.random.randint(0, 10, size=(100,))
+        from scipy.signal import chirp
+        self._audio_sample_npy = chirp(np.linspace(0, 100), 500, 2, 100)
+
+        self._text_string = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrs" \
+                            "tuvwxyz0123456789"
+
+        if tf is not None:
+            input_shape = [1, 28, 28]
+
+            l = tf.keras.layers
+            max_pool = l.MaxPooling2D(
+                (2, 2), (2, 2), padding='same', data_format="channels_first")
+            self._model_tf = tf.keras.Sequential(
+                [
+                    l.Reshape(
+                        target_shape=input_shape,
+                        input_shape=(28 * 28,)),
+                    l.Conv2D(
+                        32,
+                        5,
+                        padding='same',
+                        data_format="channels_first",
+                        activation=tf.nn.relu),
+                    max_pool,
+                    l.Conv2D(
+                        64,
+                        5,
+                        padding='same',
+                        data_format="channels_first",
+                        activation=tf.nn.relu),
+                ])
+        else:
+            self._model_tf = None
+
+        if torch is not None:
+            self._model_torch = torch.nn.Sequential(
+                torch.nn.Conv2d(3, 8, 3, padding=1),
+                torch.nn.ReLU(),
+                torch.nn.Conv2d(8, 1, 3, padding=1),
+                torch.nn.LeakyReLU(),
+                torch.nn.Conv2d(1, 23, 3),
+            )
+
+        else:
+            self._model_torch = None
+
+        self._embedding_npy = np.random.rand(500, 3)
+
+        self._labels_npy = np.random.randint(0, 10, 100)
+        self._predictions_npy = np.random.randint(0, 10, 100)
+
+        self._logger = self._setup_logger()
+
+    def _setup_logger(self):
+        return make_logger(TensorboardBackend(
+            {"log_dir": os.path.join(".", "runs", self._testMethodName)}
+        ))
+
+    @staticmethod
+    def _destroy_logger(logger: Logger):
+        logger.close()
+        del logger
+        gc.collect()
+
+    def test_image_npy(self):
+        self._logger.log({"image": {"tag": "image_npy",
+                                    "img_tensor": self._npy_imgs[0]}})
+
+    @unittest.skipIf(torch is None, "Torch Backend not installed")
+    def test_image_torch(self):
+        self._logger.log({"image": {"tag": "image_torch",
+                                    "img_tensor":
+                                        torch.from_numpy(self._npy_imgs[0])}})
+
+    def test_img_npy(self):
+        self._logger.log({"img": {"tag": "img_npy",
+                                  "img_tensor": self._npy_imgs[0]}})
+
+    @unittest.skipIf(torch is None, "Torch Backend not installed")
+    def test_img_torch(self):
[…]
