+++ delira/models/classification/classification_network.py
@@ -67,7 +67,7 @@ if "TORCH" in get_backends():
 
         @staticmethod
         def closure(model: AbstractPyTorchNetwork, data_dict: dict,
-                    optimizers: dict, losses={}, metrics={},
+                    optimizers: dict, losses=None, metrics=None,
                     fold=0, **kwargs):
             """
             closure method to do a single backpropagation step
@@ -108,6 +108,10 @@ if "TORCH" in get_backends():
 
             """
 
+            if losses is None:
+                losses = {}
+            if metrics is None:
+                metrics = {}
             assert (optimizers and losses) or not optimizers, \
                 "Criterion dict cannot be emtpy, if optimizers are passed"
 
