+++ delira/data_loading/dataset.py
@@ -583,7 +583,13 @@ class Nii3DLazyDataset(BaseLazyDataset):
         """
 
     @make_deprecated('LoadSample')
-    def __init__(self, data_path, load_fn, img_files, label_file, **load_kwargs):
+    def __init__(
+            self,
+            data_path,
+            load_fn,
+            img_files,
+            label_file,
+            **load_kwargs):
         """
          Parameters
         ----------
