+++ delira/training/predictor.py
@@ -284,8 +284,11 @@ class Predictor(object):
         """
         if metrics is None:
             metrics = {}
-        yield from self.predict_data_mgr_cache(datamgr, batchsize, metrics,
-                                               metric_keys, verbose,
+        yield from self.predict_data_mgr_cache(datamgr=datamgr,
+                                               batchsize=batchsize,
+                                               metrics=metrics,
+                                               metric_keys=metric_keys,
+                                               verbose=verbose,
                                                cache_preds=False, **kwargs)
 
         return
@@ -330,8 +333,11 @@ class Predictor(object):
         """
         if metrics is None:
             metrics = {}
-        yield from self.predict_data_mgr_cache(datamgr, batchsize, metrics,
-                                               metric_keys, verbose,
+        yield from self.predict_data_mgr_cache(datamgr=datamgr,
+                                               batchsize=batchsize,
+                                               metrics=metrics,
+                                               metric_keys=metric_keys,
+                                               verbose=verbose,
                                                cache_preds=True, **kwargs)
 
         return
@@ -382,11 +388,16 @@ class Predictor(object):
 
         if metrics is None:
             metrics = {}
+
         predictions_all, metric_vals = [], {k: [] for k in metrics.keys()}
 
-        for preds, _metric_vals in self.predict_data_mgr(datamgr, batchsize,
-                                                         metrics, metric_keys,
-                                                         verbose, **kwargs):
+        for preds, _metric_vals in self.predict_data_mgr(
+                datamgr=datamgr,
+                batchsize=batchsize,
+                metrics=metrics,
+                metric_keys=metric_keys,
+                verbose=verbose,
+                **kwargs):
             if cache_preds:
                 predictions_all.append(preds)
             for k, v in _metric_vals.items():
