+++ delira/training/predictor.py
@@ -169,7 +169,7 @@ class Predictor(object):
             a dictionary containing all metrics of the current batch
 
         """
-
+        print(metric_keys)
         if metrics is None:
             metrics = {}
         orig_num_aug_processes = datamgr.n_process_augmentation
@@ -282,6 +282,7 @@ class Predictor(object):
         does not cache anything by default
 
         """
+        print(metric_keys)
         if metrics is None:
             metrics = {}
         yield from self.predict_data_mgr_cache(datamgr=datamgr,
@@ -398,6 +399,7 @@ class Predictor(object):
                 metric_keys=metric_keys,
                 verbose=verbose,
                 **kwargs):
+
             if cache_preds:
                 predictions_all.append(preds)
             for k, v in _metric_vals.items():
