### 2019-06-15 bae7e2e
Add copy of kwargs in kfold; move sampler to Augmenter instead of dataloader to avoid multiprocessing issues
