### 2019-06-15 6e31d18
remove deprecated comment
