### 2019-06-15 41601d2
Correctly close queue
