### 2019-06-15 fa6be59
remove debug prints
