### 2019-06-15 6f53150
Add numba trafos
