### 2019-06-15 8da8abc
reduce num_processes to 2 for CI/CD
