### 2019-06-15 3542724
Prevent deadlocks
