### 2019-06-15 a2fabd1
Only apply JIT to __call__
