### 2019-06-15 9595964
Add outputs to determine hanging function
