### 2019-06-15 e63bde8
remove stopiteration in dataloader
