+++ tests/data_loading/test_data_loader.py
@@ -1,7 +1,7 @@
 import unittest
 
 import numpy as np
-from multiprocessing import Manager
+from multiprocessing import SimpleQueue
 from delira.data_loading import BaseDataLoader, SequentialSampler
 from . import DummyDataset
 
@@ -10,8 +10,7 @@ class DataLoaderTest(unittest.TestCase):
 
     def test_data_loader(self):
         np.random.seed(1)
-        manager = Manager()
-        sampler_queue = manager.Queue()
+        sampler_queue = SimpleQueue()
         dset = DummyDataset(600, [0.5, 0.3, 0.2])
         sampler = SequentialSampler.from_dataset(dset)
         loader = BaseDataLoader(dset, batch_size=16,
@@ -20,17 +19,17 @@ class DataLoaderTest(unittest.TestCase):
         sampler_queue.put(sampler(16))
 
         self.assertIsInstance(loader.generate_train_batch(), dict)
-        sampler_queue.put_nowait(sampler(16))
+        sampler_queue.put(sampler(16))
 
         for key, val in loader.generate_train_batch().items():
             self.assertEqual(len(val), 16)
 
-        sampler_queue.put_nowait(sampler(16))
+        sampler_queue.put(sampler(16))
 
         self.assertIn("label", loader.generate_train_batch())
-        sampler_queue.put_nowait(sampler(16))
+        sampler_queue.put(sampler(16))
         self.assertIn("data", loader.generate_train_batch())
-        sampler_queue.put_nowait(sampler(16))
+        sampler_queue.put(sampler(16))
 
         self.assertEqual(
             len(set([_tmp
