### 2019-06-15 d8ff501
Add chancel_join_thread
