### 2019-06-15 8937893
switch to manager with queue proxy
