### 2019-06-15 6dfaefb
enable debug mode
