### 2019-06-16 e928e31
Remove debug mode
