+++ tests/training/test_experiments.py
@@ -1,10 +1,11 @@
 
 from delira import get_backends
-from delira.utils.context_managers import DebugEnabled
 import unittest
 
 import numpy as np
 from functools import partial
+import logging
+logger = logging.getLogger(__name__)
 from delira.training import Parameters
 
 from delira.data_loading import AbstractDataset
@@ -146,7 +147,7 @@ class ExperimentTest(unittest.TestCase):
             )
 
         self._test_cases_tf = test_cases_tf
-        print(self._testMethodName)
+        logger.info(self._testMethodName)
 
     @unittest.skipIf("TORCH" not in get_backends(),
                      reason="No TORCH Backend installed")
@@ -278,23 +279,21 @@ class ExperimentTest(unittest.TestCase):
         from delira.training import TfExperiment
         from delira.data_loading import BaseDataManager
 
-        with DebugEnabled():
+        for case in self._test_cases_tf:
+            with self.subTest(case=case):
+                (params, dataset_length_train, dataset_length_test,
+                 network_cls) = case
 
-            for case in self._test_cases_tf:
-                with self.subTest(case=case):
-                    (params, dataset_length_train, dataset_length_test,
-                     network_cls) = case
+                exp = TfExperiment(params, network_cls,
+                                   key_mapping={"images": "data"})
 
-                    exp = TfExperiment(params, network_cls,
-                                       key_mapping={"images": "data"})
+                dset_train = DummyDataset(dataset_length_train)
+                dset_test = DummyDataset(dataset_length_test)
 
-                    dset_train = DummyDataset(dataset_length_train)
-                    dset_test = DummyDataset(dataset_length_test)
+                dmgr_train = BaseDataManager(dset_train, 16, 2, None)
+                dmgr_test = BaseDataManager(dset_test, 16, 1, None)
 
-                    dmgr_train = BaseDataManager(dset_train, 16, 2, None)
-                    dmgr_test = BaseDataManager(dset_test, 16, 1, None)
-
-                    exp.run(dmgr_train, dmgr_test)
+                exp.run(dmgr_train, dmgr_test)
 
     @unittest.skipIf("TF" not in get_backends(),
                      reason="No TF Backend installed")
