### 2019-06-16 9534077
Make tests only run once
