### 2019-06-16 cf5cf76
Make codestyle changes
