+++ tests/training/test_losses_torch.py
@@ -37,7 +37,7 @@ class FocalLossTestPyTorch(unittest.TestCase):
         # target for focal loss
         p_t = p * t + (1 - p) * (1 - t)
         alpha_t = torch.Tensor([alpha]).expand_as(t) * t + \
-                  (1 - t) * (1 - torch.Tensor([alpha]).expand_as(t))
+            (1 - t) * (1 - torch.Tensor([alpha]).expand_as(t))
         w = alpha_t * (1 - p_t).pow(torch.Tensor([gamma]))
         fc_value = F.binary_cross_entropy(p, t, w, reduction='none')
 
@@ -45,7 +45,7 @@ class FocalLossTestPyTorch(unittest.TestCase):
         p_tmp = torch.sigmoid(p_l)
         p_t = p_tmp * t + (1 - p_tmp) * (1 - t)
         alpha_t = torch.Tensor([alpha]).expand_as(t) * t + \
-                  (1 - t) * (1 - torch.Tensor([alpha]).expand_as(t))
+            (1 - t) * (1 - torch.Tensor([alpha]).expand_as(t))
         w = alpha_t * (1 - p_t).pow(torch.Tensor([gamma]))
 
         fc_value_logit = \
