### 2019-06-16 ee31b59
Re-add logging after merge
