+++ delira/logging/__init__.py
@@ -1,3 +1,11 @@
-from .multistream_handler import MultiStreamHandler
-from .trixi_handler import TrixiHandler, VisdomLoggingHandler, \
-    TensorboardXLoggingHandler
+from delira.logging.tensorboard_backend import TensorboardBackend
+from delira.logging.visdom_backend import VisdomBackend
+from delira.logging.base_backend import BaseBackend
+from delira.logging.writer_backend import WriterLoggingBackend
+from delira.logging.base_logger import Logger, SingleThreadedLogger, \
+    make_logger
+from delira.logging.utils import unregister_logger, register_logger, \
+    get_logger, logger_exists, log as _log
+from delira.logging.logging_context import LoggingContext
+
+log = _log
