### 2019-06-17 #1488 @haarburger
@labarba, @arfon we've finally tagged our repository and uploaded v0.4.0 to Zenodo at `10.5281/zenodo.3247307`
