### 2019-06-17 468d88b
Integrate new logging into base trainer
