+++ delira/logging/__init__.py
@@ -5,7 +5,7 @@ from delira.logging.writer_backend import WriterLoggingBackend
 from delira.logging.base_logger import Logger, SingleThreadedLogger, \
     make_logger
 from delira.logging.utils import unregister_logger, register_logger, \
-    get_logger, logger_exists, log as _log
+    get_logger, logger_exists, log as _log, get_available_loggers
 from delira.logging.logging_context import LoggingContext
 
 log = _log
