+++ delira/training/base_trainer.py
@@ -5,8 +5,8 @@ import typing
 
 import numpy as np
 from tqdm import tqdm
+from delira.logging import LoggingContext, log
 
-from delira.logging import TrixiHandler
 from .callbacks import AbstractCallback
 from .predictor import Predictor
 from ..data_loading.data_manager import Augmenter
@@ -151,6 +151,7 @@ class BaseNetworkTrainer(Predictor):
         self.stop_training = False
         self.save_freq = save_freq
         self.metric_keys = metric_keys
+        self._logger_name = None
 
         for cbck in callbacks:
             self.register_callback(cbck)
@@ -417,77 +418,84 @@ class BaseNetworkTrainer(Predictor):
 
                 val_metric_keys[k] = v
 
-        for epoch in range(self.start_epoch, num_epochs + 1):
+        with LoggingContext(self._logger_name):
 
-            self._at_epoch_begin(metrics_val, val_score_key, epoch,
-                                 num_epochs)
+            for epoch in range(self.start_epoch, num_epochs + 1):
 
-            batch_gen_train = datamgr_train.get_batchgen(seed=epoch)
+                self._at_epoch_begin(metrics_val, val_score_key, epoch,
+                                     num_epochs)
 
-            # train single network epoch
-            train_metrics, train_losses = self._train_single_epoch(
-                batch_gen_train, epoch, verbose=verbose)
+                batch_gen_train = datamgr_train.get_batchgen(seed=epoch)
 
-            total_metrics = {
-                **train_metrics,
-                **train_losses}
+                # train single network epoch
+                train_metrics, train_losses = self._train_single_epoch(
+                    batch_gen_train, epoch, verbose=verbose)
 
-            # validate network
-            if datamgr_valid is not None and (epoch % self.val_freq == 0):
-                # next must be called here because self.predict_data_mgr
-                # returns a generator (of size 1) and we want to get the first
-                # (and only) item
-                val_metrics = next(
-                    self.predict_data_mgr_cache_metrics_only(
-                        datamgr_valid, datamgr_valid.batch_size,
-                        metrics=val_metric_fns, metric_keys=val_metric_keys,
-                        verbose=verbose))
+                total_metrics = {
+                    **train_metrics,
+                    **train_losses}
 
-                total_metrics.update(val_metrics)
+                # validate network
+                if datamgr_valid is not None and (epoch % self.val_freq == 0):
+                    # next must be called here because self.predict_data_mgr
+                    # returns a generator (of size 1) and we want to get the
+                    # first (and only) item
+                    val_metrics = next(
+                        self.predict_data_mgr_cache_metrics_only(
+                            datamgr_valid, datamgr_valid.batch_size,
+                            metrics=val_metric_fns,
+                            metric_keys=val_metric_keys,
+                            verbose=verbose))
 
-            for k, v in total_metrics.items():
-                total_metrics[k] = reduce_fn(v)
+                    total_metrics.update(val_metrics)
 
-            # check if metric became better
-            if val_score_key is not None:
-                if val_score_key not in total_metrics:
-                    if "val_" + val_score_key not in total_metrics:
-                        logger.warning(
-                            "val_score_key '%s' not a valid key for \
-                                    validation metrics" % str(val_score_key))
+                for k, v in total_metrics.items():
+                    total_metrics[k] = reduce_fn(v)
 
-                        new_val_score = best_val_score
+                # check if metric became better
+                if val_score_key is not None:
+                    if val_score_key not in total_metrics:
+                        if "val_" + val_score_key not in total_metrics:
+                            logger.warning(
+                                "val_score_key '%s' not a valid key for \
+                                        validation metrics" %
+                                str(val_score_key))
[…]
